Replaced React-Table with a basic HTML table
React-Table v6 isn't compatible with v7. V7 is hooks-only. V6 is no longer supported by the original creator/maintainer. So I have decided to remove that dependency completely as it doesn't add that much to the basic workings of plugin.

Preparations for FireFox. Still buggy in FireFox so need more fixes there.